Add Previously Created Code
Add code created in Eclipse from several months ago. The code may not initially compile due to the package declaration. Run the program in a console to use. First, the program will ask for some settings to use for the encryption or decryption, then it will ask for the text to convert. The program will end by printing out the encrypted / decrypted text.